+++ src/channel.ts
@@ -170,18 +170,10 @@ export class Channel<StreamChatGenerics extends ExtendableGenerics = DefaultGene
    * @return {Promise<SendMessageAPIResponse<StreamChatGenerics>>} The Server Response
    */
   async sendMessage(message: Message<StreamChatGenerics>, options?: SendMessageOptions) {
-    const sendMessageResponse = await this.getClient().post<SendMessageAPIResponse<StreamChatGenerics>>(
-      this._channelURL() + '/message',
-      {
-        message,
-        ...options,
-      },
-    );
-
-    // Reset unreadCount to 0.
-    this.state.unreadCount = 0;
-
-    return sendMessageResponse;
+    return await this.getClient().post<SendMessageAPIResponse<StreamChatGenerics>>(this._channelURL() + '/message', {
+      message,
+      ...options,
+    });
   }
 
   sendFile(
@@ -406,7 +398,19 @@ export class Channel<StreamChatGenerics extends ExtendableGenerics = DefaultGene
       this._channelURL(),
       update,
     );
+
+    const areCapabilitiesChanged =
+      [...(data.channel.own_capabilities || [])].sort().join() !==
+      [...(Array.isArray(this.data?.own_capabilities) ? (this.data?.own_capabilities as string[]) : [])].sort().join();
     this.data = data.channel;
+    // If the capabiltities are changed, we trigger the `capabilities.changed` event.
+    if (areCapabilitiesChanged) {
+      this.getClient().dispatchEvent({
+        type: 'capabilities.changed',
+        cid: this.cid,
+        own_capabilities: data.channel.own_capabilities,
+      });
+    }
     return data;
   }
 
@@ -649,7 +653,7 @@ export class Channel<StreamChatGenerics extends ExtendableGenerics = DefaultGene
    * @see {@link https://getstream.io/chat/docs/typing_indicators/?language=js|Docs}
    * @param {string} [parent_id] set this field to `message.id` to indicate that typing event is happening in a thread
    */
-  async keystroke(parent_id?: string) {
+  async keystroke(parent_id?: string, options?: { user_id: string }) {
     if (!this.getConfig()?.typing_events) {
       return;
     }
@@ -663,6 +667,7 @@ export class Channel<StreamChatGenerics extends ExtendableGenerics = DefaultGene
       await this.sendEvent({
         type: 'typing.start',
         parent_id,
+        ...(options || {}),
       } as Event<StreamChatGenerics>);
     }
   }
@@ -672,7 +677,7 @@ export class Channel<StreamChatGenerics extends ExtendableGenerics = DefaultGene
    * @see {@link https://getstream.io/chat/docs/typing_indicators/?language=js|Docs}
    * @param {string} [parent_id] set this field to `message.id` to indicate that typing event is happening in a thread
    */
-  async stopTyping(parent_id?: string) {
+  async stopTyping(parent_id?: string, options?: { user_id: string }) {
     if (!this.getConfig()?.typing_events) {
       return;
     }
@@ -681,6 +686,7 @@ export class Channel<StreamChatGenerics extends ExtendableGenerics = DefaultGene
     await this.sendEvent({
       type: 'typing.stop',
       parent_id,
+      ...(options || {}),
     } as Event<StreamChatGenerics>);
   }
 
@@ -714,7 +720,7 @@ export class Channel<StreamChatGenerics extends ExtendableGenerics = DefaultGene
   async markRead(data: MarkReadOptions<StreamChatGenerics> = {}) {
     this._checkInitialized();
 
-    if (!this.getConfig()?.read_events) {
+    if (!this.getConfig()?.read_events && !this.getClient()._isUsingServerAuth()) {
       return Promise.resolve(null);
     }
 
@@ -732,7 +738,7 @@ export class Channel<StreamChatGenerics extends ExtendableGenerics = DefaultGene
   async markUnread(data: MarkUnreadOptions<StreamChatGenerics>) {
     this._checkInitialized();
 
-    if (!this.getConfig()?.read_events) {
+    if (!this.getConfig()?.read_events && !this.getClient()._isUsingServerAuth()) {
       return Promise.resolve(null);
     }
 
@@ -1241,7 +1247,6 @@ export class Channel<StreamChatGenerics extends ExtendableGenerics = DefaultGene
       case 'message.read':
         if (event.user?.id && event.created_at) {
           channelState.read[event.user.id] = {
-            // because in client.ts the handleEvent call that flows to this sets this `event.received_at = new Date();`
             last_read: new Date(event.created_at),
             last_read_message_id: event.last_read_message_id,
             user: event.user,
@@ -1290,6 +1295,12 @@ export class Channel<StreamChatGenerics extends ExtendableGenerics = DefaultGene
             channelState.addPinnedMessage(event.message);
           }
 
+          // do not increase the unread count - the back-end does not increase the count neither in the following cases:
+          // 1. the message is mine
+          // 2. the message is a thread reply from any user
+          const preventUnreadCountUpdate = ownMessage || isThreadMessage;
+          if (preventUnreadCountUpdate) break;
+
           if (event.user?.id) {
             for (const userId in channelState.read) {
               if (userId === event.user.id) {
@@ -1304,9 +1315,7 @@ export class Channel<StreamChatGenerics extends ExtendableGenerics = DefaultGene
             }
           }
 
-          if (ownMessage) {
-            channelState.unreadCount = 0;
-          } else if (this._countMessageAsUnread(event.message)) {
+          if (this._countMessageAsUnread(event.message)) {
             channelState.unreadCount = channelState.unreadCount + 1;
           }
         }
@@ -1360,6 +1369,23 @@ export class Channel<StreamChatGenerics extends ExtendableGenerics = DefaultGene
           delete channelState.members[event.user.id];
         }
         break;
+      case 'notification.mark_unread': {
+        const ownMessage = event.user?.id === this.getClient().user?.id;
+        if (!(ownMessage && event.user)) break;
+
+        const unreadCount = event.unread_messages ?? 0;
+
+        channelState.read[event.user.id] = {
+          first_unread_message_id: event.first_unread_message_id,
+          last_read: new Date(event.last_read_at as string),
+          last_read_message_id: event.last_read_message_id,
+          user: event.user,
+          unread_messages: unreadCount,
+        };
+
+        channelState.unreadCount = unreadCount;
+        break;
+      }
       case 'channel.updated':
         if (event.channel) {
           const isFrozenChanged = event.channel?.frozen !== undefined && event.channel.frozen !== channel.data?.frozen;